amountは小数も入力

+++ controller/schema/food.go
@@ -1,9 +1,9 @@
 package schema
 
 type Food struct {
-	ID            *int   `json:"id,omitempty"`
-	Name          string `json:"name"`
-	Amount        int    `json:"amount,omitempty"`
-	Unit          string `json:"unit"`
-	LastUpdatedAt *int64 `json:"last_updated_at,omitempty"`
+	ID            *int    `json:"id,omitempty"`
+	Name          string  `json:"name"`
+	Amount        float64 `json:"amount,omitempty"`
+	Unit          string  `json:"unit"`
+	LastUpdatedAt *int64  `json:"last_updated_at,omitempty"`
 }
