+++ ent/schema/recordfood.go
@@ -15,7 +15,7 @@ func (RecordFood) Fields() []ent.Field {
 	return []ent.Field{
 		field.Int("record_id").Positive().Unique(),
 		field.Int("food_id").Positive(),
-		field.Int("amount").Positive().Default(1),
+		field.Float("amount").Positive().Default(1),
 	}
 }
